Make shift toggle rather than momentary
It doesn't seem to be possible to have a state which is enabled when a
momentary button is disabled.

Also, the APCKEY25 sends a note off with velocity 127 when releasing the
shift button, which does not disable the state (see
StateController:handle_message which only disables the state if the
velocity is the minimum, i.e. 0).